+++ events/interactionCreate.js
@@ -1,54 +1,39 @@
 // module for interaction event of client
-// this function gets called when interacting with user, which should be with every command
-// TODO: read this when getting to interations https://discordjs.guide/creating-your-bot/creating-commands.html#replying-to-commands
-// I like the isCommand
+// this function gets called when interacting with user, which should be with every command (slash commands)
 
 const { reply } = require('../helper_functions/event_helper.js');
 
 module.exports = {
-        // AYYO DICSORD JS GUIDE KINDA SUS GOTTA USE REST API
-        // USE THIS: https://www.youtube.com/watch?v=-YxuSSG_O6g        
-	name: 'INTERACTION_CREATE',
+        // update for v14: no longer need to use API to run slash commands
+	name: 'interactionCreate',
         ws: true,
 	async execute(interaction, client) {
-                const { name, options } = interaction.data;
-                const commandName = name.toLowerCase();
-                // print the command
-                console.log(commandName);
-                // reply(client, interaction, 'pog?');
-                if (options)
-                        console.log(`options[0] are ${JSON.stringify(options[0])}`);
+                
+                if (!interaction.isChatInputCommand()) return; // if not a slash command message, ignore it
+                if (!interaction.isRepliable()) return; // if cannot reply, then don't even try!
 
-                // process this message
-                const args = [];
-                if (options) {
-                        for (const option of options) {
-                                const { value } = option;
-                                console.log(`option is ${value}`);
-                                args.push(value);
-                        }
-                }
+                const commandName = interaction.commandName.toLowerCase();
 
-                // reply(client, interaction, 'ok');
-                console.log(`interaction: ${JSON.stringify(interaction)}`);
-                console.log(`\n\n\n\n\n\n\n\n\n`); // ${JSON.stringify(interaction.message)}`);
+                // print interaction information
+                console.log(`interaction: ${interaction}`);
+                console.log(`\n\n`);
 
+                // get cooldowns to get ensure this command is not on cooldown
                 const { cooldowns, default_cooldown } = client;
-                /************************************ preprocessing of arguments ************************************/
-                // based on https://discordjs.guide/creating-your-bot/commands-with-user-input.html#basic-arguments
 
                 // if message comes from bot, ignore it
                 if (interaction.member.user.bot) return;
 
-                /************************************ actual commands ************************************/
+                /************************************ check user permissions and cooldown ************************************/
                 // https://discordjs.guide/command-handling/#dynamically-executing-commands
-                if (!client.commands.has(commandName)) return;
+                if (!client.commands.has(commandName)) return; // iif not a command, skip
                 const command = client.commands.get(commandName);
 
-                if (!command.public && interaction.member.id !== client.my_maker.id) return; // if not public and I don't ask for it, ignore
+                // if not public and I (owner) don't ask for it, ignore
+                if (!command.public && interaction.member.id !== client.my_maker.id) return;
 
-                // if admin command, get out of here!
-                if (command.admin && !interaction.member.hasPermission('ADMINISTRATOR')) return;
+                // if admin command and you aren't an admin, get out of here!
+                if (command.admin && !interaction.memberPermissions.has('ADMINISTRATOR')) return;
 
                 // she's a beaut: https://discordjs.guide/command-handling/adding-features.html#expected-command-usage
                 // if incorrectly formatted, send strongly worded message
@@ -82,9 +67,10 @@ module.exports = {
                 timestamps.set(interaction.member.id, now);
                 setTimeout(() => timestamps.delete(interaction.member.id), cooldownAmount);
 
+                /************************************ finally, run the command ************************************/
 
                 try {
-                        command.execute(interaction, args, client); // run command with args and database reference
+                        command.execute(interaction, client); // run command
                 } catch (error) { // if there's an error, print it as well as a message in the chat
                         console.error(error);
                         reply(client, interaction, 'there was an error trying to execute this command :/');